CE-163: Add support for signatures on the Link identity transaction
Now, the identity transaction requires a digital signature by one
of the linked keys to the given identity, this signature is verified
before applying the transaction to the state.

This implies that now, a key can be linked to an existing identity
only if the transaction has been signed by one of the keys that
belongs to that identity, otherwise, the transaction is rejected.