A-667: Scalanet integration with OBFT (#310)
* A-667: Create `OBFTNetworkInterface` and `OBFTChannel`

This is a draft design for an integration of scalanet to OBFT.
The code compiles and tests still pass but the application should crash with this changes.
After review, if the approach is endorsed, I will attempt to refine the implementation using a `UDPPeerGroup`.

* A-667: Address review suggestions

This commit restores the original OBFT interface. OBFT will only handle two
streams to communicate with the network. One for receiving messages from the
other servers and the second one to queue messages to diffuse to the rest of
the cluster.
A network interface was implemented to transparently manage the connection of
these two streams with the actual message communication.
Two draft implementations illustrate how to implement the network interface.
One uses scalanet `PeerGroup`s and the other one simulates the current local
communication for which we have direct access to the input streams of other servers.
The proof of concept application that simulates a cluster of nodes is now working
correctly using the network interface that implements local communication.

* A-667: Create `atala.obft.network` package

A `network` package was created and code was cleaned up.

* A-667: Refactor OBFTChannel

Now the `OBFTChannel` is the one that provides the output stream for the
`OuroborosBFT` instances instead of receiving it as a parameter.

* A-667: Integrate a UDP network based interface for OBFT

A very simple version of a UDP network interface and channel was added.
The AppPoC class is now communicating through UDP.
Pending work for future commits:
- Load network configuration settings from a file
- Improve the connection initiation between servers: Servers are currently
  using two scalanet channels per pair to communicate with each other. This
  is not a big penalty using UDP channels, but can and need to be improve to
  minimise the use of resources.

* A-667: Try to fix AppPoC and improve initialisation process for the OBFT network interface

- The network interface now only creates one channel per pair of servers.
- The channel gets updates when new servers get connected after channel initialisation.
- The AppPoC was not working on the terminal and it seems that IntelliJ was running another version of the code.
- Now AppPoC runs on the terminal but UDP connections are failling.

A-667 Fix decco problem

* A-667: Refactor network package

- The network package now lives under `atala` and at the same level as `obft` package.
- An extra StateGate class kept by git during rebase was removes.
- UDP communication on AppPoC is still not working.

* A-667: Fix UDP connections issue

Changes
- Add a few tests to the `network` package
- Fix UDP communication in AppPoC (the issue was a bad use of monix library)
- Add nicer interface to AppPoC

* A-667: Add proper logging

Code cleanup and logging in the UDP network interface package

* A-667: Address PR review comments

Changes:
- Remove unnecessary `Codecs` import in `OBFTActorMessage`
- Move network interface initialisation to `Server` class `run` method
- Close network channel on server shutdown procedure

* A-667: Address final PR comments

- Use constructor for `OBFTPeerGroupChannel` instead of `apply` method
- Rewrite tests in order to guarantee the call of `shutdown` method in network interfaces and `close` method in network channels