Disable the legacy build on circleci
It's known to the team that circleci is failing most builds, and
most of them aren't related to our changes but just to failures in
bazel while building the legacy project.

The legacy project hasn't been updated for a while, and it's unlikely
to update it for some time (at least it won't be updated before
deliveriing CVP).

Let's start getting used to have circleci in green again before merging PRs.