ATA-2256 (WIP) Add the GetWalletStatus API to the browser wallet.
This is marked as WIP as another PR will be done in order to add the
CreateWallet API. This round will help me figure out any details before
working on the next part.

This was tested manually, as no unit tests exist. I can volunteer to add
some in another ticket in order to increase coverage, which would
probably help me anyway. LMK if you'd like me to do so.