ATA-2286: Add proposal for credential verification process (#656)
* ATA-2286: Add proposal for credential verification process

This commit:
- Moves the original protocol description described as a sidetree variation on top of Bitcoin
- Creates a v0.2 of the protocol adapted for Cardano

The commit contributes with a draft of the protocol state and a proposal for the verification process for credentials

* ATA-2286: Add draft description for the protocol operations

This commit adds descriptions for the Create DID, Update DID, Issue Credential and Revoke Credential operations.

* ATA-2286: Address PR comments

The commit edits the text and expanded on required points.
It also adds:
- A section related to future changes to consider
- A section on a generic credentials' schema proposal and
  compliance notes with W3C standard drafts

* ATA-2286: Minor corrections from PR reviews

The commit just adds editorial changes, no major update is made.
The sole exception may be that we clarified that operations should
have the exact form we expect (i.e. no missing nor extra fields).
This was ambiguous in the text.

* ATA-2286: Remove unnecessary comparison with BLOCKCHAIN_TX_TIMESTAMP

This commit removes an unnecessary check done in credentials' creation and revocation
operations.
We used to have the checks

```
state.dids(issuerDIDSuffix).keys(signingKeyId).keyAdditiontionEvent < BLOCKCHAIN_TX_TIMESTAMP &&
(state.dids(issuerDIDSuffix).keys(signingKeyId).keyRevocationEvent.isEmpty ||
 state.dids(issuerDIDSuffix).keys(signingKeyId).keyRevocationEvent.get > BLOCKCHAIN_TX_TIMESTAMP )
```

However, we then realised that no event can occur after BLOCKCHAIN_TX_TIMESTAMP because there is no
way to process an operation "from the future". Note that the first proposition:
```
state.dids(issuerDIDSuffix).keys(signingKeyId).keyAdditiontionEvent < BLOCKCHAIN_TX_TIMESTAMP &&
```
was wrong because it was missing the case where the values are equal (i.e. when operations are batched)

We simplified the check to
```
state.dids(issuerDIDSuffix).keys(signingKeyId).keyRevocationEvent.isEmpty
```

* ATA-2286: Add note related to order of operations when batching is implemented