ATA-2256: Add CreateWallet API to browser wallet. (#678)
* ATA-2256: Add CreateWallet API to browser wallet.

First of all, I apologize for the big commit, ideally this would've been
broken up but I'm still figuring out the easiest way to do so in Git
(without a lot of work).

Again, this doesn't add tests but they will be added when working on
ATA-2346. Of course, it was tested manually.

To test manually, this was added to the Runner, and does a wallet status
check and then decides to either create it or simply load it. To make
this more like the desktop wallet, I added an error for when the wallet
is being loaded but doesn't exist yet.

Also, I apologize for some formatting changes that were introducing by
me being over lovely to the formatting shortcut, will try to configure
it to only format changed lines, and not all the file.

There is some piece of code that wasn't really tested: popup/Runner. I'm
not exactly sure we want this, so LMK if I should remove it.

* Resolve first round of comments:

  - Use WalletData to keep the internal state in WalletManager.
  - Log failed Futures.
  - Improve idioms.
  - Return a Future when saving in WalletManager to not "fire and
forget".

* Resolve comment round #2:

  - Better condition checking
  - println() use over "dom" logs.
  - Isolate internal state mutations.

* Simplify match's default case.