Add tests for BackgroundAPI. (#708)
* ATA-2346: Clean up of the browser wallet.

Highlights:
  - Dead code removal
  - Avoid calling the storage local when the wallet is already loaded
  - Log-to-println conversion
  - Fixes for warnings, mostly of list-to-tuple auto conversions

* Refactor out the wallet initialization in the background runner.

So it can be avoided during tests.

Another idea was to add a config but didn't feel right, also with the
development of a new PR then it makes sense to not do this automatically
anyway.

* Create fake APIs that are only available in an actual browser.

These APIs were stubbed out super simply, with the minimum features.
Note some types were set as IntelliJ was kindly asking for them,
although no actual warning was given by sbt.

Also note the fake encryption/decryption is very simple, as we only need
to ensure this step is done and not ignored as it's very important. LMK
if you see something wrong with adding a delta to a byte :)

* Add tests for BackgroundAPI.

All tests are only on the exposed API and are not testing some browser
interactions, so LMK if you'd prefer to test something different
instead, or in addition to it.

I think testing BackgroundAPI covers a lot of the functionality, and
definitely the important one. But LMK if there are other code you think
deserves its own unit tests so I can tackle them.

Also note I'm using a newer ScalaTest that has AsyncWordSpec which I
think it's very nice and avoids the need for ScalaFutures (I actually
tried to use it but had some weird deadlocks, probably because of
ScalaJS and being truly async, there's a note about this).

* Sync with develop.

Accommodate changes from the Mnemonic PR, which includes not
automatically initializing the wallet, so kinda making one of the
previous commits in this PR obsolete.

* Move key creation to popup.Runner.

So that keys are created when the wallet is created. Currently no keys
are being created and hence nothing can be signed.

* Resolve first round of comments.

From Marcin and Jacek.

* Final changes to resolve comments:

  - Add a few more tests
  - Drop HttpService as it's not used nor planned to be used
  - Few idiomatic improvements

* Delete file I missed in the previous commit.

* Add the new tests to CircleCI.

Note there are some building errors when packaging the extension, but
given it's for optional dependencies it doesn't actually fail. Example:

[error] npm WARN optional SKIPPING OPTIONAL DEPENDENCY: yallist@3.1.1 (node_modules/fsevents/node_modules/yallist):

I tried to fix them but wasn't lucky enough, so will just go with them
as it's not actual problem as it's not needed for testing as we're not
doing integration testing... yet.