ATA-2354: Expose the build info in the Node RPC service. (#720)
* Ignore the client-storage.txt file in Git.

This file gets created when (I think :P) playing with the desktop
wallet, and should not be in the repo.

* Generate a io.iohk.cvp.BuildInfo object containing build info.

I ended up using the mill.contrib.buildinfo plugin as it seemed better
for consistency with our usage of Mill. Unfortunately, it means the
generated BuildInfo object can only have strings as fields (IIUC) so we
end up doing some parsing when exposing it in the RPC (next commit).

Please note I kept it simple (as in I found the gist and simplified it)
but took the liberty to use a "standard" for our published versions,
which is generated from Git tags and current state instead of manually
curated (which we could do if we really want to control it).

This format is better explained in build.sc, but it essentially boils
down to using the tag, or
<LAST_TAG>-<N_COMMITS>-<CURRENT_COMMIT><DIRTY_SUFFIX>, which somewhat
guarantees uniqueness.

Because of the format, I didn't feel like exposing the tag and/or the
commit separately made sense, but LMK if you have better suggestions or
feel otherwise.

* Expose the build info in the Node RPC service.

Note the test cannot test the "version" properly, but this should be OK
:)

* Add the get-build-info command to the Node CLI.

Example run:

$ mill node.client.run get-build-info
version: "vAlpha-02-116-7b2603-DIRTYbf6b4bdc"
scalaVersion: "2.12.10"
millVersion: "0.6.1"
buildTime: 1583992348997

* Switch version format to <BRANCH>-<N_COMMITS>-<CURRENT_COMMIT>.

For consistency with Docker versioning. Was lucky to reuse the code
 :)