Fix test flakiness due to build time. (#745)
It turns out 10 minutes is not enough for these tests in a local
environment. CI would be fine with it because it will always rebuild (I
think?) but locally if the src code doesn't change then the binary
doesn't get rebuilt again hence the build time doesn't change, but if a
test gets executed again then it will fail if the last src build was >
10 min ago.

This fix widens the check to a month, as I'd expect to be long enough
that everyone updates their local branch in that timeframe. If you don't
feel it makes sense I could also just remove it completely, or even use
a year, as we only care that's some "date".