ATA-2411: Test JS protos for interactive demo are up to date. (#779)
This is borrowing the foundations laid out by
https://github.com/input-output-hk/atala/pull/778.

To make the protoc installation reusable, I refactored the steps out to
a command, which are now also used by a new job
build-credentials-verification-landing.

Given the are no unit tests inside -landing, I simply let the JS proto
check be the only test. For consistency, I left everything pretty
similar to -web, so in the future when there are unit tests (I hope!)
they will look the same.

Note I again had to update the protos for the new test to pass, so we
gooood.