Upgrade scalafmt to 2.5.1. (#809)
The webextension was already using it, which was inconsistent with the
rest of the projects. I upgraded the rest and unified the conf file to
also be linked to the parent conf file, so we need to only update one
for all projects next time.

Made the install-scalafmt.sh to force the installation as well.

Also removed old sbtPlugin on webextension, things seem to work but LMK
if this is wrong.