ATA-2515: Add client for Cardano Wallet API. (#807)
* ATA-2515: Add client for Cardano Wallet API.

I tested integration manually with a "test" that talked to a local
wallet and it did discover a couple problems that are now reflected in
the implementation :)

Also, note I deviated from the naming w.r.t. the Bitcoin side, as the
bitcoind API uses json/rpc whilst Cardano Wallet uses REST. LMK your
thoughts on naming and organization too.

I also think I need a proper package for Cardano DB Sync, in case you
feel funny about the new org, but will do that in another PR of course.

I'm putting all models in the same `models` package because they will
most likely make it all the way to the main future Cardano Client.
Because of this, I decided to just return the TransactionId when
creating a transaction, instead of the object, and needed the proper
decoder for that, in case you wonder why I did what I did ;)

* Resolve various comments from Christos :)

* Explain the WalletId TODO why SHA256Value cannot be extended.

* Fix small formatting issue.

* Simplify test resource path.

And remove unused val.