ATA-2265: "Process" Atala objects from Cardano transactions. (#835)
* Refactor out testing utilities for Cardano DB Sync.

This will be needed to be reused by higher level testing.

* Add a service to persist key-value pairs.

This relies on a new key_values table, but tries to hide the
implementation in case we switch to another way in the future.

* Refactor CardanoClient to support CardanoLedgerService.

Mainly to be able to query by block number (instead of hash) and remove
dead code. Also expose getLatestBlock as it's needed.

* Wait on 31 blocks to confirm Cardano transactions.

This also includes the processing of Atala objects found in confirmed
transactions, but given the lack of metadata support is not actually
done, but left ready for it.

* Pull merged https://github.com/input-output-hk/atala/pull/829

* Resolve Marcin's comments.

Clarified many things with docs. Made more idiomatic changes.