Upgrade scalapb-grpcweb to v0.3.0. (#860)
* Upgrade scalapb-grpcweb to v0.3.0.

And enable webextension tests as this release is no longer a snapshot
and should be more stable when checking it out.

I tried to upgrade all the way to version 0.4.1 but faced an error that
seems to indicate we need to upgrade to ScalaJS 1.1 before we can use
it:

Caused by: org.scalajs.ir.IRVersionNotSupportedException: This version (1.1) of Scala.js IR is not supported. Supported versions are up to 1.0

* Remove hex encoding from Mnemonic.

I think there's a mismatch about the expected typing and the real one
here, I cannot reproduce it locally though but given this method is not
needed I went with removing for now, which I think it's a far price for
having tests. For posterity, this was the error:

[error] /home/circleci/repo/credentials-verification-webextension/src/main/scala/io/iohk/atala/cvp/webextension/common/Mnemonic.scala:9:45: type mismatch;
[error]  found   : String("hex")
[error]  required: typings.node.BufferEncoding
[error]     bip39.mnemonicToSeedSync(seed).toString("hex")
[error]                                             ^

Co-authored-by: Christos KK Loverdos <christos.loverdos@iohk.io>