ATA-2608: Ask user for password on wallet recovery and registration. (#873)
* Set up Selenium testing with Chrome.

Got rid of FakeCryptoApi as it's now natively provided by Chrome.

* Add a trait to simplify DOM testing.

And use it in the only existing tests. New ones will use it too, with
actual DOM usage.

* Ask user for password for both registration and recovery.

Also ask for password verification and show any validation errors to the
user. This was extra but felt it was important to do.

Also did some changes for consistency between views.

Backfilled some tests I felt were "free". Added TODOs for the ones that
need a bigger change and will file a ticket to track them.

* Refactor form validation checks to its own util.