Improve webextension caching. (#877)
* Improve webextension caching.

Moved the extension packaging so its deps get cached as well.

.ivy2/local was also being used to download stuff to, so changed
.ivy2/cache to grab more with .ivy2.

Added coursier caching as well.

Also, scalajs-bundler was having its own node_modules just for test, I
dunno why but added it as well.

Note I also introduced a scala-version param, but didn't want to reuse
it in other places as I don't think it was semantically correct.

* Add plugins.sbt to the webextension cache key.

Also introduces the prefix "v1" as we can afford it as the key is
changing anyway. Future changes to the paths will require a bump to
"v2".