ATA-2690: Enable extra Scalac options to all CV backend. (#884)
And clean up the warnings to make it compile again.

Note that Twirl templates became a problem because they
automatically get some imports added to the generated Scala code. I had
to install the "silencer" compiler plugin in order to be able to
suppress warnings on these files. I hope it's not too broad, but we can
get back to this and tweak it.