ATA-2683: Skip CircleCI jobs when components are not affected. (#895)
* ATA-2683: Skip CircleCI jobs when components are not affected.

For simplicity, a bash script is added to detect whether a branch
deploys an image, because when it does, all tests should run to
guarantee the image quality.

For other branches, it detects all the changes since the last develop
fork (need to check whether *not* rebasing works properly) and ignores
some files (e.g., documentation).

When one of the directory parameters have been detected as changed, the
script returns a non-zero exit code. A zero return code means no
directory changed and testing can be skipped.

Note that changes outside a component directory won't be detected, but I
think this is OK? LMK otherwise.

This is all encapsulated in a new CircleCI command, which is called from
jobs that can be skipped. Because the job does run, it should become
"green" on GitHub when it's skipped.

I put the CV component as a dependency to all other components, but LMK
if you'd want me to streamline this.

Once this is merged, I will try to watch closely all runs in CircleCI to
detect any issues.

For the record, there are ways to completely skip the jobs, but we would
need to heavily modify the workflow and write a script to launch the
workflow with dynamic params depending on the affected directories. I
think this PR is way simpler and easier to understand and may get us
where we want to be.

It should be important to note that all testing will still be done to
PRs when merged to develop, and that authors should receive an email if
their PR failed to pass tests in develop, so as long develop is green,
this work is doing what it has to do :)

* Remove dep of android with backend.

Changing the backend doesn't really affect the Android side, not today,
maybe tomorrow once protos are used directly, perhaps.

* Tighten script style.

* Fix the strict issues caused in a previous commit.

Had to disable pipefail checks for `head` as it's expected :?