ATA-2761: Setup crypto subproject. (#929)
* ATA-2761: Setup crypto subproject.

This subproject will hold all crypto work that needs to be cross-compiled
to both JS and JVM (Scala). For now, only `EC` will live there, which is
basically a unified wrapper for `EC` in webextension and `ECKeys` in CV.

Several tasks, like this, have been added to ATA-2741 to split the work.
This current task is only for setting up the project and makes the
following decisions:
  - It uses the same name `EC` to implement the shared `ECTrait` in both
JS and JVM. This allows clients to also be cross-compiled and not worry
about having to know in which implementation they are running, whether
it's JS or JVM.
  - It contains a base test that should test all common functionality.
Ideally, they will be all possible tests, but we will see.
  - It uses the scalajs-bundler plugin only for JS tests, so that the
elliptic NPM dep is available. This may change for regular code too, if
we find out it's helpful on the client side when we start using it.

I discovered that the key pair encoding is different in JS and JVM, and
plan to resolve that if easy right here, but if not, I need professional
guidance :) and will implement it when adding the rest of the methods.

Feel free to ask any other question I have failed to answer already.

Be kind (NOT!), this is literally my first sbt project from scratch.

* Address review comments.