ATA-2762: Implement all EC methods. (#970)
* ATA-2762: Implement all EC methods.

There's some small refactoring that happened to improve readability. I
also implemented our encoding found in CV's ECKeys, and a decoding
(wasn't asked for, but thought it made sense to have it).

Note that the JVM signing already hashes the data with SHA256, but JS
did not so I had to add that layer in the JS side. Tests show this is
now equivalent.

Note I hit a ScalaJS bug where a negative Byte value will get
hex-encoded with "fffff", whilst the JVM impl won't. I simply fixed this
by truncating the formatting. All tests pass, which is what matters? :)

For testing, I always generate a random key-pair except in a couple cases to
guarantee both JS and JVM implementation lead to the same results.

* Resolve review comments.