ATA-2764: Replace ECKeys by Crypto's EC in the backend. (#995)
* Open up ECConfig.

So clients can know the elliptic curve used, for example. Once we
abstract things better we may close it down again, but right now is
necessary.

* Add crypto as a dep to the backend (common).

This customizes the `compile` `mill` step to publish locally the current code of Crypto and uses `latest.integration` to ensure this version is loaded as a dep.

Note it cannot track when Crypto has actually changed, so if you make
changes to it and want to see it reflected immediately after in the
backend then a `sbt publishLocal` is needed inside `crypto/`.

* Replace ECKeys by EC in ConnectorCLI.

* Replace ECKeys with EC in the Connector.

* Replace ECKeys with EC in the Node.

Note that the `poc` package uses Odissey which in turns needs Java keys
which we cannot get from the Crypto traits as they're hidden in the
impl. Because of this, that packages wasn't migrated, and a legacy
method somewhere had to be added. Once we have Odissey in our repo we
can customize it, and perform the migration, or simply remove the
package if we want to :)

* Add crypto as a dep in CircleCI.

Also install sbt in the backend CircleCI config.

* Fix bad rebase.

Well, more like, migrate the recently merged code :)