ATA-2854: Temporarily change issuance key check (#1004)
Given that the key derivation process is not fully implemented yet, the
wallet is using (for now) the master key to issue credentials. This
commit updates the node to allow this.
Originally, the node expected an issuance key both for issuing and
revoking credentials. Now, a master key is accepted for issuance and
revocation.
We must take this change back to the original state after implementing
the proper key derivation process in the wallet