ATA-2935: Add Java bindings for credential verification. (#1036)
* ATA-2935: Add Java bindings for credential verification.

The idea is to have a facade, like in
https://github.com/input-output-hk/atala/pull/1028, to implement a Java
interface which is exposed to Android. Other objects are simple POJOs,
which can become more complex over time, but for now should suffice.

Tests are similar to the Scala bare impl. We only focus on the Android
impl because we're more interested on testing the "glue" between Java
and Scala.

As usual, names might not be best, feel free to suggest better!

* Use japi.ECPublicKey instead of Scala's.