Fix directories_changed.sh script because of grep. (#1041)
I realized https://github.com/input-output-hk/atala/pull/1040 ran some
tests but it shouldn't because it's a MD only change. The issue was that
grep returns 1 when no line is selected, for example, for empty input.
See https://stackoverflow.com/a/42251542.

This is normal behavior, so only way seems to disable and then re-enable
`set -e`.