Automatically retry some tests to avoid flakiness. (#1063)
Some tests are flaky to compile because of, I think, our limited CircleCI
plan and our implicit use of medium machines. Because it doesn't fail
all the time, I hope to just be able to retry and avoid the manual
retries we need to do now, especially since all tests are required to
merge PRs.

Note that I could get rid of `&&` because CircleCI fails when any
command fails anyway.