Improve the Crypto-Backend integration. (#1077)
For ease of use, this change isolates the Crypto integration in a single
module so it can just be imported into others.

Most notably, this change:
  - Does not use the "latest" version anymore, but the current one. This
avoids issues where a previously built version that was newer was used.
  - It publishes the Crypto lib every time a mill command is ran,
meaning there's some extra overhead (few secs) but it always has the
latest code. This makes it easier to work with when modifying both
projects.
  - Uses the mill logger for logging, instead of the famous println :)
  - Fixes the `mill mill.scalalib.GenIdea/idea` issue when Crypto has
not been published before. This was because that module is implemented
in a way that does not call `compile` directly, so had to resort to
using `resolveDeps`, which gets called several times but isolating into
its own module and using `T.input` made it so it runs only once per mill command.