ATA-3013: Allow running tests within IntelliJ. (#1091)
* ATA-3013: Allow running tests within IntelliJ.

The basic idea is to hide the "scoverage" object, so everything is
dynamic and generating the IntelliJ files with `mill
mill.scalalib.GenIdea/idea` does not "see" any Scoverage dependency
because is not static.

I went with Alexis original idea of switching the behavior depending on
being on CircleCI or not, but we could remove that if it becomes an
issue somehow. Other options is to run the tests as normal without code
instrumentation, and then again but with code instrumentation. The
latter is actually a better practice, but given we don't want to run
these tests for so long then we will just go this way for now.

To test this, checkout the branch and run `mill mill.scalalib.GenIdea/idea` and then restart IntelliJ invalidating the cache. Tests should work now, ping me if they don't.

* Now that instrumentation does not happen at the earliest compile step,
try to get rid of the split to simplify our commands.