Avoid failing CircleCI when trigger-integration-tests fails. (#1104)
Because of Browserstack, the trigger-integration-tests step in our CircleCI workflow has been failing in develop since forever. Unfortunately, this is hiding real issues as we get emailed every time it fails, which is always, so when another step fails, it's hard to tell without examining the email whether it needs our attention or not.

Just like any other test that is permanently broken or is out of the developers power to fix, we should disable it until it's supposed to work.

I hope I don't come too strongly on this, but it really feels like the
right thing to do. I do try to pay a lot of attention to CircleCI, so
this really gets in the way of doing that easily and properly.

This commit should be very easy to revert, in order to re-enable failing
and fixing https://jira.iohk.io/browse/ATA-3042.