Simplify code with AtalaObjectNotification. (#1142)
* Simplify code with AtalaObjectNotification.

Yes, I'm not super excited with the name, but the best ones were already
taken and `AtalaReference` did not feel right because the `update` could
be either a `Reference` or a `ByteContent`. Feel free to suggest
something better, as always.

This spawned many changes, which I think leads to naming consistency.

And this is all coming from
https://github.com/input-output-hk/atala/pull/1127#discussion_r485952387

If you have something better in mind, or this is just too overboard, of
course LMK.

* Fix compilation error.

Duh.

* Disable CardanoClientExample test as it's currently failing.

We will investigate the issue but for now let's unblock the team by
disabling it.