ATA-2511: Cardano tx metadata format design for PRISM. (#1165)
* ATA-2511: Cardano tx metadata format design for PRISM.

Went with a simple design that I somewhat shared with a couple people,
which seems enough to get us started. The main thing is to allow us to
evolve the schema as necessary, which is taken into account.

Updated the README to talk about the things done and point to the schema
design.

* Change schema to store content as an array of integers.

Instead of a bytestring because they are limited.

Also resolve comments, too.

* Simplify the schema even further.

By stating `content` is the proto bytes and we can evolve that instead.

Introduce PRISM_INDEX too.

* Make `version` is own field.

Instead of being implicit as a key.