ATA-2516: Read metadata from cardano-db-sync. (#1192)
* Read tx metadata from DB Sync.

DB Sync stores the top-level keys as `key` in the `tx_metadata` table,
so we need to build a higher level JSON object when we read from DB
Sync.

Note only PRISM metadata will be returned. This was not done merely as
an optimization (although it is), but to ease parsing the JSON because
of the issue mentioned above about `key` being originally part of the
JSON too.

* Notify of incoming Atala objects in Cardano.

* Add AtalaObjectMetadata to encode and decode AtalaObjects to/from JSON.

This was done by TransactionMetadata but wasn't the best place as it was
meant to be more generic than that.

Having this code here also allows for more thorough testing, which
actually discovered a couple silly bugs.