Retry webextension tests on CircleCI. (#1244)
The tests take ~1 min to run so a retry is not expensive and would let
us avoid having to re-run the workflow.

Why does it fail? It's an old problem I added when I moved it to test on
Chrome, so I think it's lack of resources which we could probably only
fix by upgrading our CircleCI machine plan. The issue is definitely a
framework/infra problem, so nothing we can do, unfortunately.