ATA-3427: Decouple transaction info from AtalaObject. (#1272)
Now that's decoupled, we also store the AtalaObject when publishing, and
simply populate the transaction info when found in the blockchain (we
still add the AtalaObject if not stored before anyway).

The decoupling in the DB really comes from the fact that it was cleaner
than having a bunch of optional fields/columns.

As a plus, we've started storing the block number as well, which was
defaulted to `1` for all past transactions. This is not a problem as we
haven't launched yet and we would only use this data for informational
purposes only, for now.