ATA-3451: Implement getting and deleting transactions for all ledgers. (#1283)
In order to retry, we need to get and delete transactions, and while this has been implemented for Cardano, it has not for the other ledgers and we need to provide these methods in `AtalaReferenceLedger` in order for `ObjectManagementService` to use it.

For now, we won't implement these methods for Bitcoin, and will only assume transactions will always make it to the blockchain. This is just to save time because we don't need to support Bitcoin at the moment, and will probably need extensive testing to revive it anyway.