ATA-3353: Retry old pending transactions. (#1287)
* ATA-3352: Record every AtalaObject published to the ledger.

A new table was created which only keeps history for now. Next PR will
take this information and be able to retry.

I added the status field and getBy filter as it was somewhat needed for
testing, and will be used in the next PR anyway.

Note I also removed one test that I needed to update but then realized
it wasn't necessary to have so I just killed it.

* ATA-3353: Retry old pending transactions.

It sets a default 12 min timeout to declare a transaction as "rolled
back" ("stuck", "not accepted"), which is roughly 31 Cardano blocks. I
am yet to make another change so this value becomes a no-op for other
ledgers, but it's just a small refactoring that prefer to tackle in a
different PR.

* Fix bad merge.

* Resolve comments.