ATA-3537: Implement a new metadata schema for Cardano. (#1333)
* ATA-3537: Implement a new metadata schema for Cardano.

Based on the analysis done in https://github.com/input-output-hk/atala/blob/c5fb2115879d53e5a636d1ee503766349dc24951/credentials-verification/docs/cardano/metadata-schema-fee.md

We're breaking compatibility with the previous schema, as it's not worth
to maintain it because we haven't released. And because we're changing
the "version" key to "v" we can keep using "1" as the version number.

Note that with this we uncovered something we ignored: cardano-wallet
decodes bytestrings from a string starting with `0x` whereas
cardano-db-sync stores bytestrings as an object with a `hex` key. I
discovered this during integration testing, which has our backs :)

With this change, and the reason given above,
encoding/decoding a `TransactionMetadata` object is no longer idempotent
:/ and so some tests had to change, plus some changes to improve
manteinance for new schemas.

* Resolve comments and Scala update.

Also log the transaction ID used in integration testing because
sometimes we want to debug.