ATA-3736: Fix Merkle Tree scalacheck generators (#1420)
* ATA-3736: Remove unnecessary toList invocations

* ATA-3736: Make hashNonEmptyListGen generate distinct hashes

Non-distinct generator leads to situations when it does not matter whether
a node is a right sibling or a left one. In turn, this makes one of the tests
fail (albeit rarely) as incorrect Merkle Tree index does not lead into an
invalid proof. Anyway, Merkle Trees usually operate on distinct hashes so I
do not think shrinking our test case base is a big deal here.