Always aggregate test results in CircleCI. (#1572)
When tests fail, the aggregation steps in CircleCI do not run, and hence
when it tries to upload the test results it does not find them because
they are scattered in multiple subdirectories. This means we can never
see the failing test without looking at the logs.

This change ensure that both test results and the coverage data is
aggregated and hence uploaded. Note steps `store_test_results` and
`store_artifacts` always run, so no need to change that, only our
customs ones need to.

Also, added a retry for the SDK JS tests, as we know, they are flaky and
a simple retry will make our builds greener.