ATA-4074: Update Cardano metadata JSON schema. (#1575)
Basically the idea was to make it compatible with the newest
cardano-wallet and cardano-db-sync. These Cardano components were
manually updated, but a PR will be coming later to do it properly.

The changes needed for cardano-wallet was to annotate the metadata with
their types (you can see how weird map became) and for cardano-db-sync
some simplification was done and some column renaming as well.

I will update the schema cost analysis in another PR as well.

The Cardano integration test is also being re-enabled here.