ATA-3025: SQL schema for "credential issuances". (#1521)
* ATA-3025: SQL schema for "credential issuances".

I agree it's a weird name, but it really is the best I could come up
with (well, Alex did), because:
  - "Credential batches" was gonna be confusing with the "Slayer
batches", or perhaps we should re-brand those to "Slayer batches"?
Although it's a bit late.
  - "Credential groups" was also gonna be confusing with "issuer groups"
("contact groups").
  - "Issuance" is what we actually want here, so it does make sense.

This schema is based on the needs for the mockups
(https://www.figma.com/file/N27WwOG5GxhQmoAoKDzARe/ATALA-_-Management-Console?node-id=2584%3A0, under Flow: Credential Management), and a bit future proof. Main features are:
  - Credential issuances can be left as a draft, and finished later.
  - We do distinguish between finishing uploading the data and having
all resulting credentials published on the blockchain.
  - It does end up generating normalized `credentials` rows, but with a
new field so we can trace back, and potentially filter appropriately
depending on the screen (list of issuances vs. list of credentials).
These rows are created until the very end, until then data can be
changed.
  - Either groups of contacts or direct contacts can be added to the
credential issuance.
  - Groups are copied over from `issuer_groups`, as a way to freeze them
and allow changing the underlying groups without affecting the issuance.
Ideally, we will realize if a change exists and ask the user in the
frontend whether to update the frozen groups or not.
  - `credential_issuance_contacts` links directly to `contacts` as a
reference, because `credential_data` is what matters to it.
  - An issuance may be soft deleted in the future, but there won't be a
UI for it and hence no DB status either.

* Resolve comments.

* Move SQL schema to MC (as it's a new feature).

Add more indices.

* Fix script for MC.